Update deleteAllEntities implementation in dataDelete.js
Use recursion

+++ sashimi-webapp/src/database/data-modifier/dataDelete.js
@@ -9,9 +9,18 @@ const sqlCommands = new SqlCommands();
 export default class dataDelete {
   static constructor() {}
 
-  static deleteAllEntities() {
-    for (let index = 0; index < constants.ARRAY_ENTITIES_NAME.length; index += 1) {
-      sqlCommands.deleteTable(constants.ARRAY_ENTITIES_NAME[index]);
+  static deleteAllEntities(index) {
+    if (typeof Promise === 'function') {
+      return new Promise((resolve, reject) => {
+        if (index < constants.ARRAY_ENTITIES_NAME.length -1) {
+          sqlCommands.deleteTable(constants.ARRAY_ENTITIES_NAME[index])
+          .then(data => this.deleteAllEntities(index+1))
+          .catch(sqlError => reject(sqlError));
+          resolve(true);
+        }
+      });
+    } else {
+      throw new exceptions.PromiseFunctionNotDefined();
     }
   }
